refactor(backoffice-v2): created dumb cells for dialog and call to action, approve actions refactor
