+++ apps/backoffice-v2/src/pages/Entity/components/CallToActionLegacy/interfaces.ts
@@ -1,0 +1,18 @@
+import { ICallToActionDocumentSelection } from '@/pages/Entity/components/DirectorsCallToAction/interfaces';
+
+export interface ICallToActionLegacyProps {
+  value: {
+    text: string;
+    props: {
+      id: string;
+      disabled: boolean;
+      decision: 'reject' | 'approve' | 'revision' | 'revised';
+      documentSelection?: ICallToActionDocumentSelection;
+      contextUpdateMethod?: 'base' | 'director';
+      revisionReasons?: string[];
+      rejectionReasons?: string[];
+      onReuploadReset?: () => void;
+      onDialogClose?: () => void;
+    };
+  };
+}
