+++ apps/backoffice-v2/src/pages/Entity/components/CaseCallToActionLegacy/interfaces.ts
@@ -1,7 +1,7 @@
 import { TWorkflowById } from '../../../../domains/workflows/fetchers';
 import { CommonWorkflowStates } from '@ballerine/common';
 
-export interface ICaseCallToActionProps {
+export interface ICaseCallToActionLegacyProps {
   value: string;
   data: {
     parentWorkflowId: string;
@@ -9,8 +9,8 @@ export interface ICaseCallToActionProps {
     childWorkflowContextSchema: TWorkflowById['childWorkflows'][number]['workflowDefinition']['contextSchema'];
     disabled: boolean;
     decision:
-      | CommonWorkflowStates.REJECTED
-      | CommonWorkflowStates.APPROVED
-      | CommonWorkflowStates.REVISION;
+      | typeof CommonWorkflowStates.REJECTED
+      | typeof CommonWorkflowStates.APPROVED
+      | typeof CommonWorkflowStates.REVISION;
   };
 }
