+++ apps/backoffice-v2/src/pages/Entity/components/DirectorsCallToAction/interfaces.ts
@@ -1,0 +1,10 @@
+export interface ICallToActionDocumentOption {
+  name: string;
+  value: string;
+}
+
+export interface ICallToActionDocumentSelection {
+  options: ICallToActionDocumentOption[];
+  value?: ICallToActionDocumentOption['value'];
+  onSelect: (value: ICallToActionDocumentOption['value']) => void;
+}
